+++ Server/app.js
@@ -5,8 +5,27 @@ const cookieParser = require('cookie-parser');
 const logger = require('morgan');
 const cors = require('cors');
 const app = express();
-const swaggerUi = require('swagger-ui-express'),
-swaggerDocument = require('./swagger.json');
+
+const { graphqlHTTP } = require('express-graphql');
+const { buildSchema } = require('graphql');
+
+// GraphQL Schema
+const schema = buildSchema(`
+    type Query {
+        message: String
+    }
+`);
+
+// Root Resolver
+const root = {
+    message: () => 'Hello World'
+}
+
+app.use('/graphql', graphqlHTTP({
+  schema: schema,
+  rootValue: root,
+  graphiql: true
+}));
 
 app.use(logger('dev'));
 app.use(express.json());
@@ -19,7 +38,7 @@ app.use((req, res, next) => {
   res.header("Access-Control-Allow-Methods", "GET, POST, HEAD, OPTIONS, PUT, PATCH, DELETE");
   res.header("Access-Control-Allow-Headers","Origin, X-Requested-With, Content-Type, Accept");
   next();
-})
+});
 
 const productsRoute = require('./routes/products');
 const usersRoute = require('./routes/users');
@@ -27,6 +46,7 @@ const ordersRoute = require('./routes/orders');
 const authRoute = require('./routes/auth');
 const categoryRoute = require('./routes/category');
 const homeRoute = require('./routes/home');
+const { truncate } = require('fs');
 
 app.use('/api', homeRoute);
 app.use('/api/products', productsRoute);
@@ -34,7 +54,6 @@ app.use('/api/categories', categoryRoute);
 app.use('/api/users', usersRoute);
 app.use('/api/orders', ordersRoute);
 app.use('/api/auth', authRoute);
-app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
