+++ Server/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "echo \"Error:  no test specified\" && exit 1",
     "start": "node app.js"
   },
   "engines": {
@@ -17,13 +17,14 @@
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "express-graphql": "^0.12.0",
     "express-validator": "^6.6.1",
+    "graphql": "^15.5.0",
     "jade": "^1.11.0",
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.10.0",
     "multer": "^1.4.2",
     "mysqli": "^3.0.11",
-    "nodemon": "^2.0.6",
-    "swagger-ui-express": "^4.1.5"
+    "nodemon": "^2.0.6"
   }
 }
